requirements.txt
# Python Requirements for JARVIS Brain
# Install with: pip3 install -r requirements.txt

pyserial>=3.5
SpeechRecognition>=3.10.0
pyaudio>=0.2.13
pyttsx3>=2.90

# Optional: For better voice recognition
# google-api-python-client>=2.0.0

